+++ tests/integration/models/reddit/test_submission.py
@@ -67,6 +67,14 @@ class TestSubmission(IntegrationTest):
                 'TestSubmission.test_hide'):
             Submission(self.reddit, '4b1tfm').hide()
 
+    def test_hide_multiple(self):
+        self.reddit.read_only = False
+        submissions = [Submission(self.reddit, 'fewoh'),
+                       Submission(self.reddit, 'c625v')]
+        with self.recorder.use_cassette(
+                'TestSubmission.test_hide__multiple'):
+            Submission(self.reddit, '1eipl7').hide(submissions)
+
     def test_invalid_attribute(self):
         with self.recorder.use_cassette(
                 'TestSubmission.test_invalid_attribute'):
@@ -103,6 +111,14 @@ class TestSubmission(IntegrationTest):
         with self.recorder.use_cassette(
                 'TestSubmission.test_unhide'):
             Submission(self.reddit, '4b1tfm').unhide()
+
+    def test_unhide_multiple(self):
+        self.reddit.read_only = False
+        submissions = [Submission(self.reddit, 'fewoh'),
+                       Submission(self.reddit, 'c625v')]
+        with self.recorder.use_cassette(
+                'TestSubmission.test_unhide__multiple'):
+            Submission(self.reddit, '1eipl7').unhide(submissions)
 
     def test_unsave(self):
         self.reddit.read_only = False